Moved skeleton to its own file for increased readability
